+++ test/dummy/config/initializers/allow_trycloudflare_hosts.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-if Rails.env.local?
+if Rails.env.development?
   Rails.application.config.hosts << ".trycloudflare.com"
 end